requirements.txt
streamlit
tensorflow
opencv-python-headless
pillow
pandas
matplotlib